backend/requirements.txt
anthropic>=0.25.0
voyageai>=0.2.3
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
numpy>=1.26.0
pydantic>=2.0.0
requests>=2.31.0